Delete guidelines.md
Guidelines is now broken up and moved into a folder.  The folder contains separate md pages for cognitive engines and adapters now. In the future, we will add a page on aggregators and probably one on the real-time framework. Sidebars have been updated for this change, but sidebars are still missing the new Data section.

@charliemantle