VTN-14864 Add examples for engine outputs (and link to validation contract schemas) (#60)
* Add draft json-schema for audio-fingerprinting

* Add constraints to audio fingerprinting start/stop times

* Add placeholder for master vtn-standard schema

* Move audio-fingerprint schema to schemas folder

* VTN-14864 Clean up structured-data engine output example

Added a placeholder for the schema definition

* VTN-14864 Add geolocation engine output example

Added a placeholder for the schema definition and TODO information in
the README

* VTN-14864 Add object detection engine output example

Added a placeholder for the schema definition and TODO information in
the README

* Remove libraryEntity from the master series object example types

It was already removed from the non-series object options.  This just
syncs it up.

* VTN-14864 Clean up logo detection engine output example

Adds a placeholder schema and a TODO in the README
Also extends the master vtn-standard example to account for logos

* VTN-14864 Clean up transcription engine output example

Adds a placeholder schema and a TODO in the README

* VTN-14864 Clean up speaker separation engine output example

Adds a placeholder schema and TODOs in the README

* VTN-14864 Add license plate recognition engine output example

Adds a placeholder schema and TODOs in the README

* VTN-14864 Add placeholder text validation contract

And link it up to text extraction and recognition examples

* VTN-14864 Clean up cognitive engine category docs

Reorganize and sync up the sidebar
Delete empty documents

* VTN-14864 Clean up additional cognitive engine category docs

* VTN-14864 Add validation contract for face

* VTN-14864 Make engine manifest conform to the trailing slash pattern

* VTN-14864 Add docsify plugin to support badges indicating support levels

* VTN-14864 Conform engine output and message type docs to trailing slash pattern

* VTN-14864 Add reference to vtn-standard in text extraction doc

* VTN-14864 Clean up audio fingerprinting docs

With this update, the audio-fingerprinting are now the gold-standard
example of cognitive engine category-specific docs.