VTN-24034 Proposal for adding confidence to content classification (#103)
Also formalizes and documents the use of the @id key. Since the absence
of documentation looked odd next to this new key.

Admittedly, the scalar that is usually used here is not representative
of "confidence." It would be better described as "score" or "weight."
But "confidence" has so far been our best way of expressing "fuzziness"
in AI classifications. In the future, we could still add "weight" or
something to differentiate between confidence (certainty) and weight
(relative value). But for now, I feel confidence should suffice as our
single 0-100% "grayness" metric.