VTN-20345 Linkify GraphQL queries to GraphiQL (#111)
* [WIP] VTN-20345 Linkify all inline GraphQL queries to open in GraphiQL

This assumes that any inline code snippet marked as GraphQL syntax is a
full valid GraphQL query to run in GraphiQL. But there are plenty of
areas in the docs where we have GraphQL schema syntax or incomplete or
annotated GraphQL "pseudocode" that is NOT runnable in GraphiQL but
will still receive this "Try in GraphiQL" link.

TODO: The GraphiQL base points to US prod.  Need to make it configurable
off whitelisted config values.

* [WIP] VTN-20345 Add tests for GraphiQL links

TODO: env-specific API base URL

* VTN-20345 Make API root URI environment-specific

By adding a config injection script to the build process.
It works slightly differently when developing locally using the docsify
CLI because that bypasses the build step.